move automation button and trigger to its own file
