+++ src-web/tableDefinitions/utils.js
@@ -10,14 +10,11 @@ import _ from 'lodash'
 import config from '../../server/lib/shared/config'
 import {
   AcmTable,
-  AcmLaunchLink
 } from '@open-cluster-management/ui-components'
 import {
   Label,
   LabelGroup,
-  Button,
   Tooltip,
-  Spinner,
 } from '@patternfly/react-core'
 import StatusField from '../components/common/StatusField'
 import {
@@ -30,8 +27,7 @@ import msgs from '../nls/platform.properties'
 import TruncateText from '../components/common/TruncateText'
 import { LocaleContext } from '../components/common/LocaleContext'
 import purifyReactNode from '../utils/PurifyReactNode'
-import { POLICY_AUTOMATIONS } from '../utils/client/queries'
-import { Query } from '@apollo/client/react/components'
+import AutomationButton from '../components/common/AutomationButton'
 
 // use console.log(JSON.stringify(result, circular())) to test return result from transform
 export const transform = (items, def, locale) => {
@@ -66,7 +62,7 @@ export const transform = (items, def, locale) => {
 }
 
 // use console.log(JSON.stringify(result, circular())) to test return result from transform
-export const transformNew = (items, def, locale, automationOnClick) => {
+export const transformNew = (items, def, locale) => {
   // Create column data for parent table and expandable (child) tables
   const columns = {
     colParent: [],
@@ -96,7 +92,7 @@ export const transformNew = (items, def, locale, automationOnClick) => {
     rowParent: [],
     rowChild: [],
   }
-  pushRows(items, rows, def, locale, automationOnClick)
+  pushRows(items, rows, def, locale)
   // Specify a default sortBy object for the table if it doesn't exist
   const sortBy = def.sortBy ? def.sortBy : { direction: 'asc' }
   // The index can either be an integer or a string matching the column label
@@ -144,7 +140,7 @@ export const transformNew = (items, def, locale, automationOnClick) => {
   }
 }
 
-function pushRows(items, rows, def, locale, onClickAutomation) {
+function pushRows(items, rows, def, locale) {
   let subUid = 0, expandable
   items.forEach((item, index) => {
     expandable = false
@@ -165,16 +161,6 @@ function pushRows(items, rows, def, locale, onClickAutomation) {
         value =  msgs.get(key.resourceKey, locale)
       } else if (key.type === 'boolean') {
         value = value ? true : false
-      } else if (key.label === 'automation') {
-        // Leverage the defined transformFunction to render content and store the raw value as metadata
-        value = {
-          title: key.transformFunction(item, locale, {
-            onClickAutomation: (data) => {
-              onClickAutomation(data)
-            }
-          }),
-          rawData: value
-        }
       } else if (key.transformFunction && typeof key.transformFunction === 'function') {
         // Leverage the defined transformFunction to render content and store the raw value as metadata
         value = {
@@ -415,49 +401,9 @@ export function formatAnnotationString(policy, annotationKey){
   return '-'
 }
 
-export function getAutomationLink(item, locale, args) {
+export function getAutomationLink(item, locale) {
   return (
-    <Query query={POLICY_AUTOMATIONS} variables={{ namespace: item.metadata.namespace }}>
-    {( result ) => {
-      const { data={policyAutomations: []} } = result
-      const { loading } = result
-      let found = false
-      let automationName = ''
-      data.policyAutomations.forEach((automation) => {
-        if (automation.spec && automation.spec.policyRef === item.metadata.name) {
-          found = true
-          automationName = automation.metadata.name
-        }
-      })
-      if (loading) {
-        return <Spinner size='md' />
-      }
-      if (found) {
-        return <AcmLaunchLink links={[
-          {
-              id: `automationButton-${automationName}`,
-              text: <TruncateText maxCharacters={20} text={automationName} />,
-              onClick: () => {
-                args.onClickAutomation(item)
-              },
-              label: true,
-          },
-        ]}></AcmLaunchLink>
-      }
-      return (
-        <Button
-          component="a"
-          variant="link"
-          className="automationButton"
-          onClick= {() => {
-            args.onClickAutomation(item)
-          }}
-        >
-          {msgs.get('table.actions.automation.configure', locale)}
-        </Button>
-      )
-    }}
-    </Query>
+    <AutomationButton item={item} locale={locale}></AutomationButton>
   )
 }
